Move keys_for_quantity() to .reporting.util
