+++ ixmp/reporting/util.py
@@ -1,7 +1,8 @@
-from functools import lru_cache
+from functools import lru_cache, partial
 from typing import Dict
 
 import pandas as pd
+from genno import Key
 
 #: Dimensions to rename when extracting raw data from Scenario objects.
 #: Mapping from Scenario dimension name -> preferred dimension name.
@@ -31,6 +32,40 @@ def dims_for_qty(data):
 
     # Rename dimensions
     return [RENAME_DIMS.get(d, d) for d in dims]
+
+
+def keys_for_quantity(ix_type, name, scenario):
+    """Return keys for *name* in *scenario*."""
+    from .computations import data_for_quantity
+
+    # Retrieve names of the indices of the ixmp item, without loading the data
+    dims = dims_for_qty(scenario.idx_names(name))
+
+    # Column for retrieving data
+    column = "value" if ix_type == "par" else "lvl"
+
+    # A computation to retrieve the data
+    result = [
+        (
+            Key(name, dims),
+            partial(data_for_quantity, ix_type, name, column),
+            "scenario",
+            "config",
+        )
+    ]
+
+    # Add the marginal values at full resolution, but no aggregates
+    if ix_type == "equ":
+        result.append(
+            (
+                Key(f"{name}-margin", dims),
+                partial(data_for_quantity, ix_type, name, "mrg"),
+                "scenario",
+                "config",
+            )
+        )
+
+    return result
 
 
 @lru_cache(1)
